UI Testing - Kupo
KPG-103 hopes to start building the basic concept for the main menu UI. In this update I have added a background and three styled buttons with text.
- Start Game
- Options
- Exit Game

The text style will change and in the future I would like to add three bricks in the bottom right corner and animated clouds floating across the background to make things feel more active.

The button styles are bound to change as well but was testing <3

Co-Authored-By: PLURninjaDev <177242212+PLURninjaDev@users.noreply.github.com>
Co-Authored-By: PLURninja <18559080+PLURninja@users.noreply.github.com>